orders/pyproject.toml
[tool.poetry]
name = "order-service"
version = "0.1.0"
description = ""
authors = ["roshan.bhattarai"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
alembic = "^1.8.1"
mysql-connector-python = "^8.0.30"
opentelemetry-distro = "^0.34b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
